requirements.txt
google-search-results
serpapi
BeautifulSoup4
pandas
requests
google.generativeai
python-dotenv
nltk
matplotlib
textblob
crewai
crewai[tools]



